\name{SLmix-package}
\alias{SLmix-package}
\alias{SLmix}
\docType{package}
\title{ SLMix package
\packagetitle{ SLMix packageSLmix}
}
\description{
\packageDescription{SLmix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SLmix}
\packageIndices{SLmix}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SLmix}

Maintainer: \packageMaintainer{SLmix}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
